Refactoring code, moving to separate files.
